+++ GraphicsEngine/graphics2d/graphics2d.h
@@ -41,7 +41,7 @@ int g2d_fill_triangle (
 	const int x3, const int y3);
 
 int g2d_fill_triangle_boundingbox (
-	const int x1, const int y1, 
-	const int x2, const int y2,
-	const int x3, const int y3);
+	const short x1, const short y1, 
+	const short x2, const short y2,
+	const short x3, const short y3);
 #endif
